Main program fix for number logic error
Made some changes to the main java program to fix a logic error where all contacts had the same number printed for them. This was done by using the same cursor that is used to scroll through all contact data to retrieve the contact name and contact number, rather than the code before where a separate cursor was made to retrieve numbers, separate from the overall contact one. The contacts that are scrolled through are still checked for a phone number. If the contact has a phone number, it will be printed. If the contact lacks a phone number, it will not be printed.