Added select menu and calling
Added code to select contacts in the main view. When a contact is selected, a new context menu is displayed with the options call, text, and change pokemon. Currently, only the call functionality works. In addition, it brings up the user selected contact's number in the keypad so the user may decide to go through with the call or cancel. This makes for a better user experience.