Added random number for avatar
Added retrieval of a random number in the cursor_android_contact's for loop and set the random number for each android contact there is in the phone. This helps in preventing a bug where scrolling through the contact list would change the avatar every if the contact was out of sight, then seen again. This was fixed by generating a random number for each contact and setting it. If the contact already had a random number set, return and don't set a new one. If one was not set, set the one being sent in.