Added functionality for avatar number
Added functionality for setting and getting the random avatar number for each contact. Each contact will start with an avatar number of 0 (not a possible random number.) If their avatarNum is still 0, it hasn't been set yet, so set it. If it has been set, just return and don't reset it. This prevents a scrolling bug where the avatar was changed if scrolled out of sight and then back. The number is retrieved in the PhoneBookAdapter code to set the avatar.