Added functionality for save/recover states
Added functionality to save and recover states of the app, mostly in the form of the randomly generated or selected pokemon avatars for the app owner. This was done with several methods. For changing screen orientation/restarting the app, onSaveInstanceState method was overridden and used. That method would save a hashmap that maps contact phone numbers to pokedex number to recover that contact's pokemon when restarting the app. The save data is recovered by checking for the saveInstanceState bundle being null or not during onCreate. If it's not null, the app has been used before and the hashmap will be recovered. During the main while loop that harvests all the contacts, the hashmap will be checked for each contact's number from the phone. If it's in the map, get the pokedex number corresponding to that contact number. Additionally, onStop and onResume are used for saving/recovering data when pressing the back button and coming back, etc. This is done by saving the hashmap, like in saveInstanceState, to an Intent object, and recovering it in onResume. Functionality for saving data when killing the app, most likely in the form of overriding onDestroy, will have to be done next.