Added file write/read for data recovery
Overrode the onPause method to write the savedAvatarNums hashmap to a binary file that is saved in internal storage. Then, if savedInstanceState is null, the file is attempted to be open to recover the saved data, if it exists. This is done as the data is lost when the application is killed by the system/user or the phone is restarted altogether. Now, for simple ways of leaving the app, like the back button, and for unintended or harder ways of leaving the app, like the app being killed, the data will be saved persistently for the user.