Added code for vertical alphabet usage
Added code for the vertical alphabet to be used for quickly navigating contacts. It functions by holding the index of the contact who's first letter begins with a new letter compared to previous contacts in a map. When a letter on the vertical alphabet is clicked, the clicklistener for that letter, each letter of the alphabet has its own, the listener will look in the map for the index of the contact who's name begins with that letter that appears first (contacts are in alphabetical order.) If a contact fulfills those requirements, the list will automatically jump to that contact, using setSelectionFromTop, for easy navigation.